Use Agora RtcEngine instead of UI Kit. Temp mvp
The UI kit only allows for the usage of either temp tokens
or standing up a token server to auth users.

Neither of these solutions are great, so we have to switch
to the more feature-rich rtc engine. This will give us
more flexiblitly in the long run but will be more effort
to get back the basic view.

In the Agora Video Call widget created no buttons are overlaid
which we'll need to end the call properly.