Identify client with gRPC metadata and use that to properly detect disconnects
The end/cancelled/status callbacks don't give you the ability to
call getPeer() to determine who you were connected to (not to
mention its error prone to make unique keys out of ips/ports).

The solution is on the initial connection with the server
to save off the uid of the caller and make it available
in the cancelled callback.

Will ultimately need to address the issue of any
client phone being able to identify themselves with
a random UID, impersonating another phone.