Send join request response to called; fix no transaction id
Disconnect button wasn't working because the called phone
was sending "" for the transaction id and thus the server
couldn't process it. This happened because we never sent
down the transaction id in the join request response, therefore
the called had no idea what the ID was.

As a side note, we shouldn't rely on the phones sending us this
information as it is error prone and could be exploited by spoofing
ids. Todo