EA-81 Fix call ending after 5 minutes
Implemented ping/pong support because it seemed like some grpc
libs would kill the connection after no activity, however just w/these
changes it was still happening.

Turns out cloud run had a max request timeout of 300 seconds, so by bumping
that to the max of 3600 seconds (1 hr) this issue is resolved.

Keeping the ping/pong in for extra reliablity in the connection.